refactor(tidy3d): FXC-4683-move-config-to-new-common-submodule
