+++ .github/workflows/tidy3d-python-client-tests.yml
@@ -270,6 +270,27 @@ jobs:
         run: |
           mypy --config-file=pyproject.toml
 
+  ensure-common-imports:
+    name: ensure-common-imports
+    needs: determine-test-scope
+    if: needs.determine-test-scope.outputs.code_quality_tests == 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 1
+          submodules: false
+          persist-credentials: false
+
+      - name: set-python-3.10
+        uses: actions/setup-python@v4
+        with:
+          python-version: '3.10'
+
+      - name: Run tidy3d._common import check
+        run: |
+          python scripts/ensure_imports_from_common.py
+
   zizmor:
     name: Run zizmor 🌈
     runs-on: ubuntu-latest
@@ -983,6 +1004,7 @@ jobs:
       - remote-tests
       - lint
       - mypy
+      - ensure-common-imports
       - verify-schema-change
       - lint-commit-messages
       - lint-branch-name
@@ -1003,6 +1025,12 @@ jobs:
         if: ${{ needs.determine-test-scope.outputs.code_quality_tests == 'true' && needs.mypy.result != 'success' && needs.mypy.result != 'skipped' }}
         run: |
           echo "❌ Mypy type checking failed."
+          exit 1
+
+      - name: check-common-imports-result
+        if: ${{ needs.determine-test-scope.outputs.code_quality_tests == 'true' && needs.ensure-common-imports.result != 'success' && needs.ensure-common-imports.result != 'skipped' }}
+        run: |
+          echo "❌ tidy3d._common import check failed."
           exit 1
           
       - name: check-schema-change-verification
