+++ tidy3d/_common/config/README.md
