+++ tidy3d/compat.py
@@ -1,26 +1,13 @@
-"""Compatibility layer for handling differences between package versions."""
+"""Compatibility shim for :mod:`tidy3d._common.compat`."""
 
+# ruff: noqa: F401 - ignore unused imports, imports ensure compatibility
+
+# marked as migrated to _common
 from __future__ import annotations
 
-import importlib
-from functools import cache
-
-from packaging.version import parse
-
-try:
-    from xarray.structure import alignment
-except ImportError:
-    from xarray.core import alignment
-
-try:
-    from typing import Self, TypeAlias  # Python >= 3.11
-except ImportError:  # Python <3.11
-    from typing_extensions import Self, TypeAlias
-
-
-@cache
-def _package_is_older_than(package: str, version: str) -> bool:
-    return parse(importlib.metadata.version(package)) < parse(version)
-
-
-__all__ = ["Self", "TypeAlias", "_package_is_older_than", "alignment"]
+from tidy3d._common.compat import (
+    Self,
+    TypeAlias,
+    _package_is_older_than,
+    alignment,
+)
