+++ tidy3d/config/__init__.py
@@ -1,69 +1,32 @@
-"""Tidy3D configuration system public API."""
+"""Compatibility shim for :mod:`tidy3d._common.config`."""
 
+# ruff: noqa: F401 - ignore unused imports, imports ensure compatibility
+
+# marked as migrated to _common
 from __future__ import annotations
 
-from typing import Any
+import tidy3d._common.config as _common_config
+from tidy3d.config import sections
 
-from . import sections  # noqa: F401 - ensure builtin sections register
-from .legacy import LegacyConfigWrapper, LegacyEnvironment, LegacyEnvironmentConfig
-from .manager import ConfigManager
-from .registry import (
+_common_config.initialize_env()
+
+from tidy3d._common.config import (  # noqa: E402 - import after Env setup
+    ConfigManager,
+    Env,
+    Environment,
+    EnvironmentConfig,
+    LegacyConfigWrapper,
+    LegacyEnvironment,
+    LegacyEnvironmentConfig,
+    _base_manager,
+    _config_wrapper,
+    _create_manager,
+    config,
     get_handlers,
+    get_manager,
     get_sections,
     register_handler,
     register_plugin,
     register_section,
+    reload_config,
 )
-
-__all__ = [
-    "ConfigManager",
-    "Env",
-    "Environment",
-    "EnvironmentConfig",
-    "config",
-    "get_handlers",
-    "get_sections",
-    "register_handler",
-    "register_plugin",
-    "register_section",
-]
-
-
-def _create_manager() -> ConfigManager:
-    return ConfigManager()
-
-
-_base_manager = _create_manager()
-# TODO(FXC-3827): Drop LegacyConfigWrapper once legacy accessors are removed in Tidy3D 2.12.
-_config_wrapper = LegacyConfigWrapper(_base_manager)
-config = _config_wrapper
-
-# TODO(FXC-3827): Remove legacy Env exports after deprecation window (planned 2.12).
-Environment = LegacyEnvironment
-EnvironmentConfig = LegacyEnvironmentConfig
-Env = LegacyEnvironment(_base_manager)
-
-
-def reload_config(*, profile: str | None = None) -> LegacyConfigWrapper:
-    """Recreate the global configuration manager (primarily for tests)."""
-
-    global _base_manager, Env
-    if _base_manager is not None:
-        try:
-            _base_manager.apply_web_env({})
-        except AttributeError:
-            pass
-    _base_manager = ConfigManager(profile=profile)
-    _config_wrapper.reset_manager(_base_manager)
-    Env.reset_manager(_base_manager)
-    return _config_wrapper
-
-
-def get_manager() -> ConfigManager:
-    """Return the underlying configuration manager instance."""
-
-    return _base_manager
-
-
-def __getattr__(name: str) -> Any:
-    return getattr(config, name)
